Added new Builder function in expr package
A new function named 'Builder' has been added to the 'expr' package. This function takes an expression from the 'ast' package and returns a function that operates on a build.Expr object. It uses a switch statement to handle different types of expressions, specifically 'Text' and 'VarRef'. The import statement was also updated to include the 'ast' package from 'go-make'.