Builder application (#48)
* WIP on inserting nodes

* Updated rule builder logic

Enhanced the rule builder logic in the code. The import section now includes a new package for token handling. Adjusted the AtRule function to handle nil nodes and empty targets differently, with an added position variable for better tracking. Also modified the Target method in 'nodes.go' to check if targets exist before adding space.

* Refactor builder function names for clarity

The commit includes a refactor of the builder function names in the codebase. The previous naming convention, which used 'Builder' as a suffix (e.g., RuleBuilder, ExprBuilder, FileBuilder), has been replaced with 'Func' (e.g., RuleFunc, ExprFunc, FileFunc). This change improves readability and better reflects the functional nature of these elements.

* Fiddlin

* Refactor rule transformation logic

The code changes refactor the way rules are transformed in the builder. The transformations from function 'f' are now applied directly to the node, and then the node is rewritten into a new rule to fix positions. Additionally, an enhancement has been made in copyRule function where it now handles both Text and VarRef types under Targets.

* Added instability warning to API documentation

A warning message has been added to the README file, indicating that the API is not yet stable. This serves as a caution for users interacting with the `builder` package.

* Refactored builder package structure and types

The builder package has been restructured for better organization. The contract.go file was renamed and moved to a new sub-package, build. Import paths have been updated across multiple files to reflect this change.

Type aliases were introduced in the main builder package for File, Rule, and Expr from the new build sub-package. This allows for cleaner type definitions throughout the codebase.

Function signatures were also updated to use these new types instead of old function types (ExprFunc, RuleFunc, etc.). This results in more readable and maintainable code as it's easier to understand what each function is doing.

* Refactor build.go for clarity and extensibility

The build.go file has been refactored to improve code readability and extendability. The import statements have been updated to include specific packages, enhancing the structure of the code. Additionally, method signatures in interfaces have been modified for better clarity by adding parameter names. This change aims to make the code more understandable and easier to maintain in the future.

* Added new Builder function in expr package

A new function named 'Builder' has been added to the 'expr' package. This function takes an expression from the 'ast' package and returns a function that operates on a build.Expr object. It uses a switch statement to handle different types of expressions, specifically 'Text' and 'VarRef'. The import statement was also updated to include the 'ast' package from 'go-make'.

* Refactor build.go to simplify dependencies

Removed the dependency on 'ast' package and the 'Add' method from the File interface. This simplifies the code structure and reduces unnecessary coupling, making it easier to maintain in future.

* Enhanced test coverage for rule builder

Added a new test case in the rule builder tests to verify the functionality of adding a target with a different start. This ensures that multiple targets can be added correctly and their positions are tracked accurately.

* Renamed and refactored 'mut' to 'apply'

The 'mut' module has been renamed to 'apply'. Correspondingly, the function AtRule has been renamed to ApplyRule. The test cases have also been updated to reflect these changes. This renaming provides a more accurate description of the functionality provided by this module and its methods.